pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PyA3EDA"
version = "0.0.1"
description = "Python Automatization of Automated Analysis of Electronic Structure Data"
authors = [
    { name = "Markus-G-S-Weiss", email = "markus.guenter.weiss@rwth-aachen.de" },
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "pyyaml",
]

[project.scripts]
pya3eda = "PyA3EDA.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/PyA3EDA"]
